Debugging of resource key adding
This is an example of added resource key in the translation file that is
not working correctly (nothing from the Slovak translation will be
used). I commit the changes in hope that another pair of eyes will be
able to find the problem. Note that the .json files need to be grunted
to load the changes to language packs.